SPR1-1493 Upgrade delay tracking via l0.KRETRACK
Introduce a new calibration stream ('l0') and delay calibration product
('KRETRACK'). The 'l0' stream represents tweaks to the raw data before
L1 and L2 calibration. The 'KRETRACK' product aims to redo the delay
tracking, by undoing the delays that were applied to the F-engine and
replacing them with predictions from an improved delay model.

This product only has a correction sensor because it is already split
across inputs. It makes little sense to reassemble these into a product
sensor first, only to be split again afterwards.