SPR1-1177 New extract_scan_alignment script
This utility script extracts the label, target and scan sensors of the
given dataset, both before and after scan alignment. These are printed
in tabular form. This collects data for unit tests of the align_scans
function.

Tweak MVF2 and MVF3 formats to match MVF4 by caching the raw label
sensor as the "obs_label" sensor. This allows the utility script to
work on older formats too (although the "before" and "after" versions
are currently identical because scan alignment still happens in place,
so that the original sensors are clobbered).